greedy-deps-tests: allow use of sstate for prepopulation
The only things we *need* built from scratch are the recipes we'll be testing
(target possibly with deps), so any additional bits we want to use to
prepopulate sysroot is only done to affect those recipes, so those bits are
free to come from sstate. This should make use of the -p argument less painful
in some cases, for example, if one attempts -p world.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>