leaky-layer-tests: dynamically determine targets
Rather than running bitbake -S on a particular target, instead gather up
a list of appends in the layer, run it on those targets, and then compare the
content for those targets with bitbake-whatchanged. This avoids the need to
hardcode a target, which potentially misses things, and instead tests only
what needs testing.

Signed-off-by: Christopher Larson <kergoth@gmail.com>